+++ packages/authhero/CHANGELOG.md
@@ -1,5 +1,17 @@
 # authhero
 
+## 0.59.0
+
+### Minor Changes
+
+- Create temporary tables
+
+## 0.58.0
+
+### Minor Changes
+
+- Add refresh_tokens route
+
 ## 0.57.0
 
 ### Minor Changes